Start the next release cycle
Maven only allows one single release version to be deployed, so let's
prepare for the next changes by switching to a snapshot version.

The snapshot version is in line with SemVer: http://semver.org/. In
short, SemVer suggests to

- increase the micro version for backwards-compatible bug fixes
- increase the minor version for backwards-compatible API changes
- increase the major version for backwards-incompatible changes

It is the rule that is followed by the ImageJ2 project, and more and
more also by Fiji projects.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>