Remove definite integral as it only adds redundancy to PCAs.
The idea of adding AUC was to provided an integrated measurement
of the sampled data. However, it is effectively just the average
number (mean inters.) multiplied by a constant. The goal of so
many metrics is to allow users to perform PCA. Adding redundant
variables is just superfluous clutter, and so it is best to offer
a curated repertoire of metrics not just _any_ possible metric.